sua loi phan tech, employee dang nhap duoc vao admin page

+++ website/views.py
@@ -16,10 +16,14 @@ def home(request):
 
 
 def watch_users(request):
-    if request.user.is_authenticated:
-        users = User.objects.all()
-        # Check to see if logging in
-        return render(request, 'user/user.html', {"users": users})
+    if request.user.is_authenticated :
+        if request.user.is_admin:
+            users = User.objects.all()
+            # Check to see if logging in
+            return render(request, 'user/user.html', {"users": users})
+        else:
+            messages.success(request, "You do not have permission to access this page!")
+            return redirect('employee')
     else:
         messages.success(request, "You must be Logged In to View That Page!!!")
         return redirect('login')
@@ -28,8 +32,12 @@ def watch_users(request):
 def customer_user(request, pk):
     if request.user.is_authenticated:
         # Look Up Record
-        cus_user = User.objects.get(id=pk)
-        return render(request, 'user/user_info.html', {'cus_user': cus_user})
+        if request.user.is_admin:
+            cus_user = User.objects.get(id=pk)
+            return render(request, 'user/user_info.html', {'cus_user': cus_user})
+        else:
+            messages.success(request, "You do not have permission to access this page!")
+            return redirect('employee')
     else:
         messages.success(request, "You must be Logged In to View That Page!!!")
         return redirect('login')
@@ -37,10 +45,14 @@ def customer_user(request, pk):
 
 def delete_user(request, pk):
     if request.user.is_authenticated:
-        delete_cus_user = User.objects.get(id=pk)
-        delete_cus_user.delete()
-        messages.success(request, f"User Deleted Successfully ...")
-        return redirect('users')
+        if request.user.is_admin:
+            delete_cus_user = User.objects.get(id=pk)
+            delete_cus_user.delete()
+            messages.success(request, f"User Deleted Successfully ...")
+            return redirect('users')
+        else:
+            messages.success(request, "You do not have permission to access this page!")
+            return redirect('employee')
     else:
         messages.success(request, "You must be Logged In to do that!!!")
         return redirect('login')
@@ -48,32 +60,36 @@ def delete_user(request, pk):
 
 def update_user(request, pk):
     if request.user.is_authenticated:
-        current_user = User.objects.get(id=pk)
-        if request.method == 'POST':
-            form = UpdateUserForm(request.POST, instance=current_user)
-            # Kiểm tra giá trị của username có trùng lặp hay không
-            if form.is_valid() and form.clean_username():
-                role = form.cleaned_data.get('role')
-                current_user = form.save(commit=False)
-                if role == 'is_admin':
-                    current_user.is_admin = True
-                    current_user.is_employee = False
-                    current_user.is_technician = False
-                elif role == 'is_employee':
-                    current_user.is_admin = False
-                    current_user.is_employee = True
-                    current_user.is_technician = False
-                elif role == 'is_technician':
-                    current_user.is_admin = False
-                    current_user.is_employee = False
-                    current_user.is_technician = True
+        if request.user.is_admin:
+            current_user = User.objects.get(id=pk)
+            if request.method == 'POST':
+                form = UpdateUserForm(request.POST, instance=current_user)
+                # Kiểm tra giá trị của username có trùng lặp hay không
+                if form.is_valid() and form.clean_username():
+                    role = form.cleaned_data.get('role')
+                    current_user = form.save(commit=False)
+                    if role == 'is_admin':
+                        current_user.is_admin = True
+                        current_user.is_employee = False
+                        current_user.is_technician = False
+                    elif role == 'is_employee':
+                        current_user.is_admin = False
+                        current_user.is_employee = True
+                        current_user.is_technician = False
+                    elif role == 'is_technician':
+                        current_user.is_admin = False
+                        current_user.is_employee = False
+                        current_user.is_technician = True
 
-                current_user.save()
-                messages.success(request, f'User Has Been Updated!!!')
-                return redirect('user_info', pk=current_user.id)
+                    current_user.save()
+                    messages.success(request, f'User Has Been Updated!!!')
+                    return redirect('user_info', pk=current_user.id)
+            else:
+                form = UpdateUserForm(instance=current_user)
+            return render(request, 'user/update_user.html', {'form': form, 'user_id': current_user.id})
         else:
-            form = UpdateUserForm(instance=current_user)
-        return render(request, 'user/update_user.html', {'form': form, 'user_id': current_user.id})
+            messages.success(request, "You do not have permission to access this page!")
+            return redirect('employee')
     else:
         messages.success(request, f"You must be logged in....")
         return redirect('login')
@@ -112,27 +128,34 @@ def logout_user(request):
 
 
 def register_user(request):
-    if request.method == "POST":
-        form = SignUpForm(request.POST)
-        if form.is_valid():
-            user = form.save(commit=False)
-            role = form.cleaned_data.get('role')  # Lấy giá trị của trường role từ form
-            if role == 'is_admin':
-                user.is_admin = True
-            elif role == 'is_employee':
-                user.is_employee = True
-            elif role == 'is_technician':
-                user.is_technician = True
-            user.save()
-            username = form.cleaned_data['username']
-            password = form.cleaned_data['password1']
-            user = authenticate(username=username, password=password)
-            messages.success(request, "You Have Successfully Add New User!")
-            return redirect('users')
+    if request.user.is_authenticated:
+        if request.user.is_admin:
+            if request.method == "POST":
+                form = SignUpForm(request.POST)
+                if form.is_valid():
+                    user = form.save(commit=False)
+                    role = form.cleaned_data.get('role')  # Lấy giá trị của trường role từ form
+                    if role == 'is_admin':
+                        user.is_admin = True
+                    elif role == 'is_employee':
+                        user.is_employee = True
+                    elif role == 'is_technician':
+                        user.is_technician = True
+                    user.save()
+                    username = form.cleaned_data['username']
+                    password = form.cleaned_data['password1']
+                    user = authenticate(username=username, password=password)
+                    messages.success(request, "You Have Successfully Add New User!")
+                    return redirect('users')
+            else:
+                form = SignUpForm()
+                return render(request, 'user/register.html', {'form': form})
+        else:
+            messages.success(request, "You do not have permission to access this page!")
+            return redirect('employee')
     else:
-        form = SignUpForm()
-        return render(request, 'user/register.html', {'form': form})
-    return render(request, 'user/register.html', {'form': form})
+        messages.success(request, f"You must be logged in....")
+        return redirect('login')
 
 
 def admin(request):
